+++ packages/gafl-webapp-service/src/handlers/__tests__/agreed-handler-happy-path.spec.js
@@ -1,5 +1,5 @@
 import { govUkPayApi, salesApi } from '@defra-fish/connectors-lib'
-import { initialize, injectWithCookies, start, stop } from '../../__mocks__/test-utils-system'
+import { initialize, injectWithCookies, start, stop, mockSalesApi } from '../../__mocks__/test-utils-system'
 
 import {
   ADULT_FULL_1_DAY_LICENCE,
@@ -12,11 +12,6 @@ import {
 
 import { COMPLETION_STATUS } from '../../constants.js'
 import { AGREED, TEST_TRANSACTION, TEST_STATUS, ORDER_COMPLETE } from '../../uri.js'
-
-import mockPermits from '../../__mocks__/data/permits.js'
-import mockPermitsConcessions from '../../__mocks__/data/permit-concessions.js'
-import mockConcessions from '../../__mocks__/data/concessions.js'
-import mockDefraCountries from '../../__mocks__/data/defra-country.js'
 import { PAYMENT_JOURNAL_STATUS_CODES } from '@defra-fish/business-rules-lib'
 
 beforeAll(() => {
@@ -34,10 +29,7 @@ afterAll(() => {
 })
 
 jest.mock('@defra-fish/connectors-lib')
-salesApi.permits.getAll = jest.fn(async () => new Promise(resolve => resolve(mockPermits)))
-salesApi.permitConcessions.getAll = jest.fn(async () => new Promise(resolve => resolve(mockPermitsConcessions)))
-salesApi.concessions.getAll = jest.fn(async () => new Promise(resolve => resolve(mockConcessions)))
-salesApi.countries.getAll = jest.fn(async () => new Promise(resolve => resolve(mockDefraCountries)))
+mockSalesApi()
 
 const paymentStatusSuccess = cost => ({
   amount: cost,