+++ packages/gafl-webapp-service/src/pages/licence-details/licence-length/route.js
@@ -1,6 +1,8 @@
 import { LICENCE_LENGTH, CONTROLLER } from '../../../uri.js'
 import pageRoute from '../../../routes/page-route.js'
 import Joi from '@hapi/joi'
+import { pricingDetail } from '../../../processors/pricing-summary.js'
+import { licenceTypeDisplay } from '../../../processors/licence-type-display.js'
 
 const validator = Joi.object({
   'licence-length': Joi.string()
@@ -8,4 +10,10 @@ const validator = Joi.object({
     .required()
 }).options({ abortEarly: false, allowUnknown: true })
 
-export default pageRoute(LICENCE_LENGTH.page, LICENCE_LENGTH.uri, validator, CONTROLLER.uri)
+const getData = async request => {
+  const permission = await request.cache().helpers.transaction.getCurrentPermission()
+  const pricing = await pricingDetail(LICENCE_LENGTH.page, request)
+  return { pricing, licenceTypeStr: licenceTypeDisplay(permission) }
+}
+
+export default pageRoute(LICENCE_LENGTH.page, LICENCE_LENGTH.uri, validator, CONTROLLER.uri, getData)