+++ packages/gafl-webapp-service/src/pages/order-complete/route.js
@@ -1,8 +1,8 @@
 import pageRoute from '../../routes/page-route.js'
 
+import Boom from '@hapi/boom'
 import { COMPLETION_STATUS, FEEDBACK_URI_DEFAULT } from '../../constants.js'
 import { ORDER_COMPLETE, CONTROLLER, NEW_TRANSACTION, REFUND_POLICY, ORDER_COMPLETE_PDF } from '../../uri.js'
-import Boom from '@hapi/boom'
 import { displayStartTime, displayEndTime } from '../../processors/date-and-time-display.js'
 import * as mappings from '../../processors/mapping-constants.js'
 import * as concessionHelper from '../../processors/concession-helper.js'
@@ -36,7 +36,7 @@ const getData = async request => {
     permission,
     startTimeString,
     endTimeString,
-    disabled: permission.concessions ? permission.concessions.find(c => c.type === mappings.CONCESSION.DISABLED) : null,
+    disabled: concessionHelper.hasDisabled(permission),
     licenceTypes: mappings.LICENCE_TYPE,
     hasJunior: concessionHelper.hasJunior(permission),
     hasSenior: concessionHelper.hasSenior(permission),